+++ lifelib/projects/nestedlife/nestedlife.py
@@ -83,17 +83,17 @@ def build(load_saved=False):
         refs=policy_refs)
 
     # ------------------------------------------------------------------------
-    # Build Assumptions space
+    # Build Assumption space
 
     asmp_refs = {'Policy': policy,
                  'ProductSpec': input.ProductSpec,
                  'MortalityTables': input.MortalityTables,
-                 'asmp': input.Assumptions,
+                 'asmp': input.Assumption,
                  'asmp_tbl': input.AssumptionTables}
 
     def asmp_params(PolicyID):
         refs = {'pol': Policy[PolicyID]}
-        alias = {'prd': refs['pol'].Product,
+        alias = {'prod': refs['pol'].Product,
                  'polt': refs['pol'].PolicyType,
                  'gen': refs['pol'].Gen}
         refs.update(alias)
@@ -101,15 +101,15 @@ def build(load_saved=False):
                 'refs': refs}
 
     asmp = model.import_module(
-        module_='assumptions',
-        name='Assumptions',
+        module_='assumption',
+        name='Assumption',
         formula=asmp_params,
         refs=asmp_refs)
 
     asmp.allow_none = True
 
     # ------------------------------------------------------------------------
-    # Build Assumptions space
+    # Build Assumption space
 
     def econ_params(ScenID):
         refs = {'Scenario': Input.Scenarios[ScenID]}
@@ -130,8 +130,8 @@ def build(load_saved=False):
     # 
     # lifelib --+
     #           +--BaseProjection
-    #           +--OuterProjection[PolicyID] <--- BaseProjection
-    #                    +--InnerProjection[t] <-- BaseProjection
+    #           +--OuterProj[PolicyID] <--- BaseProjection
+    #                    +--InnerProj[t] <-- BaseProjection
 
     proj_refs = {'Pol': policy,
                  'Asmp': asmp,
@@ -145,7 +145,7 @@ def build(load_saved=False):
                 'refs': refs}
 
     pvmixin = model.import_module(
-        module_='present_values',
+        module_='present_value',
         name='PV_Mixin')
 
     baseproj = model.import_module(
@@ -156,7 +156,7 @@ def build(load_saved=False):
 
     outerproj = model.new_space(
         bases=baseproj,
-        name='OuterProjection',
+        name='OuterProj',
         formula=proj_params,
         refs=proj_refs)
 
@@ -172,7 +172,7 @@ def build(load_saved=False):
 
     innerproj = outerproj.new_space(
         bases=baseproj,
-        name='InnerProjection',
+        name='InnerProj',
         formula=innerproj_params)
 
     return model