+++ tests/data/generate_testdata_nestedlife.py
@@ -16,28 +16,28 @@ def SurrRateMult(t):
         return SurrRateMult(t - 1)
 
 
-def nop_Surrender(t):
+def PolsSurr(t):
     """Number of policies: Surrender"""
-    return nop_BoP1(t) * asmp.SurrRate(t) * SurrRateMult(t)
+    return PolsIF_Beg1(t) * asmp.SurrRate(t) * SurrRateMult(t)
 
 
-def nop_EoP_inner(t):
+def PolsIF_End_inner(t):
     """Number of policies: End of period"""
     if t == t0:
-        return outer.nop_EoP(t)
+        return outer.PolsIF_End(t)
     else:
-        return nop_BoP1(t - 1) - nop_Death(t - 1) - nop_Surrender(t - 1)
+        return PolsIF_Beg1(t - 1) - PolsDeath(t - 1) - PolsSurr(t - 1)
 
 
 def set_model(model):
     # Policy point ID and aliases
     polid = 171
-    outer = model.OuterProjection[polid]
-    inner = outer.InnerProjection
+    outer = model.OuterProj[polid]
+    inner = outer.InnerProj
 
     model.BaseProjection.new_cells(formula=SurrRateMult)
-    model.BaseProjection.new_cells(formula=nop_Surrender)
-    inner.new_cells(name='nop_EoP', formula=nop_EoP_inner)
+    model.BaseProjection.new_cells(formula=PolsSurr)
+    inner.new_cells(name='PolsIF_End', formula=PolsIF_End_inner)
 
     outer.SurrRateMult[1] = 2
     outer.SurrRateMult[2] = 0.5
@@ -53,8 +53,8 @@ def set_model(model):
 def update_model(model):
 
     polid = 171
-    outer = model.OuterProjection[polid]
-    inner = outer.InnerProjection
+    outer = model.OuterProj[polid]
+    inner = outer.InnerProj
 
     outer.SurrRateMult[1] = 0.5
     outer.SurrRateMult[2] = 2
@@ -80,7 +80,7 @@ def get_nested(outer, item):
             if t < t0:
                 expect_t0[t] = 0
             else:
-                cells = outer.InnerProjection[t0].cells[item]
+                cells = outer.InnerProj[t0].cells[item]
                 expect_t0[t] = cells[t]
 
         expect.append(expect_t0)
@@ -93,7 +93,7 @@ def get_nested(outer, item):
 
 def save_data(outer, filename):
 
-    data = get_nested(outer, 'nop_Surrender')
+    data = get_nested(outer, 'PolsSurr')
     filepath = os.path.join(os.path.dirname(__file__), filename)
 
     with open(filepath, 'wb') as file: